move logic to a hook
